+++ test/dummy/app/models/payment.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+class Payment < ApplicationRecord
+  belongs_to :order
+  has_many   :refunds
+end