+++ test/dummy/app/models/product.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class Product < ApplicationRecord
+  belongs_to :category
+  has_many   :variants
+  has_many   :line_items
+  has_many   :reviews
+  has_many   :attachments, as: :attachable
+end